Moved @Provides method of FireStoreClient to FirebaseModule Removed unnecessary binds() to feature_abbreviation module as it only needs FireStoreClient
