+++ feature_abbreviations/src/main/kotlin/com/jpaya/dynamicfeatures/abbreviations/ui/di/AbbreviationsModule.kt
@@ -19,14 +19,11 @@ package com.jpaya.dynamicfeatures.abbreviations.ui.di
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.VisibleForTesting.PRIVATE
 import androidx.lifecycle.viewModelScope
-import com.google.firebase.firestore.FirebaseFirestore
 import com.jpaya.base.ui.extensions.viewModel
-import com.jpaya.englishisfun.firestore.FireStoreProperties
 import com.jpaya.dynamicfeatures.abbreviations.ui.AbbreviationsListFragment
 import com.jpaya.dynamicfeatures.abbreviations.ui.AbbreviationsListViewModel
 import com.jpaya.dynamicfeatures.abbreviations.ui.adapter.AbbreviationsListAdapter
 import com.jpaya.englishisfun.firestore.FireStoreClient
-import com.jpaya.englishisfun.firestore.FireStoreClientImpl
 import com.jpaya.dynamicfeatures.abbreviations.ui.paging.AbbreviationsPageDataSource
 import com.jpaya.dynamicfeatures.abbreviations.ui.paging.AbbreviationsPageDataSourceFactory
 import dagger.Module
@@ -58,18 +55,6 @@ class AbbreviationsModule(
     ) = fragment.viewModel {
         AbbreviationsListViewModel(dataFactory)
     }
-
-    /**
-     * Create a provider method binding for [FireStoreClient].
-     *
-     * @return Instance of data source.
-     * @see Provides
-     */
-    @Provides
-    fun providesFireStoreClient(
-        fireStore: FirebaseFirestore,
-        properties: FireStoreProperties
-    ): FireStoreClient = FireStoreClientImpl(fireStore, properties)
 
     /**
      * Create a provider method binding for [AbbreviationsPageDataSource].
