+++ feature_abbreviations/src/test/kotlin/com/jpaya/dynamicfeatures/abbreviations/ui/di/AbbreviationsModuleTest.kt
@@ -17,19 +17,12 @@
 package com.jpaya.dynamicfeatures.abbreviations.ui.di
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.google.firebase.firestore.FirebaseFirestore
-import com.jpaya.englishisfun.firestore.FireStoreProperties
 import com.jpaya.base.ui.extensions.viewModel
 import com.jpaya.dynamicfeatures.abbreviations.ui.AbbreviationsListFragment
 import com.jpaya.dynamicfeatures.abbreviations.ui.AbbreviationsListViewModel
-import com.jpaya.englishisfun.firestore.FireStoreClientImpl
 import com.jpaya.dynamicfeatures.abbreviations.ui.paging.AbbreviationsPageDataSourceFactory
 import io.mockk.*
 import io.mockk.impl.annotations.MockK
-import kotlinx.coroutines.CoroutineScope
-import org.hamcrest.MatcherAssert.assertThat
-import org.hamcrest.core.IsInstanceOf.instanceOf
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
@@ -68,33 +61,21 @@ class AbbreviationsModuleTest {
         assertEquals(dataFactory, factoryCaptor.captured().dataSourceFactory)
     }
 
-    @Test
-    fun verifyProvidedFireStoreClient() {
-        val fireStore = mockk<FirebaseFirestore>(relaxed = true)
-        val fireStoreProperties = mockk<FireStoreProperties>(relaxed = true)
-
-        val fireStoreClient = module.providesFireStoreClient(fireStore, fireStoreProperties)
-
-        assertThat(fireStoreClient, instanceOf(FireStoreClientImpl::class.java))
-        assertEquals(fireStore, (fireStoreClient as FireStoreClientImpl).fireStore)
-        assertEquals(fireStoreProperties, fireStoreClient.properties)
-    }
-
-    @Test
-    fun verifyProvidedCharactersPageDataSource() {
-        val fireStoreClient = mockk<FireStoreClientImpl>(relaxed = true)
-        val viewModel = mockk<AbbreviationsListViewModel>(relaxed = true)
-        val scope = mockk<CoroutineScope>()
-        every { viewModel.viewModelScope } returns scope
-
-        val dataSource = module.providesAbbreviationsPageDataSource(
-            fireStoreClient = fireStoreClient,
-            viewModel = viewModel
-        )
-
-        assertEquals(fireStoreClient, dataSource.fireStoreClient)
-        assertEquals(scope, dataSource.scope)
-    }
+//    @Test
+//    fun verifyProvidedCharactersPageDataSource() {
+//        val fireStoreClient = mockk<FireStoreClient>(relaxed = true)
+//        val viewModel = mockk<AbbreviationsListViewModel>(relaxed = true)
+//        val scope = mockk<CoroutineScope>()
+//        every { viewModel.viewModelScope } returns scope
+//
+//        val dataSource = module.providesAbbreviationsPageDataSource(
+//            fireStoreClient = fireStoreClient,
+//            viewModel = viewModel
+//        )
+//
+//        assertEquals(fireStoreClient, dataSource.fireStoreClient)
+//        assertEquals(scope, dataSource.scope)
+//    }
 
     @Test
     fun verifyProvidedAbbreviationsListAdapter() {
