+++ feature_abbreviations/src/test/kotlin/com/jpaya/dynamicfeatures/abbreviations/ui/paging/AbbreviationsPageDataSourceFactoryTest.kt
@@ -16,89 +16,77 @@
 
 package com.jpaya.dynamicfeatures.abbreviations.ui.paging
 
-import androidx.lifecycle.MutableLiveData
-import com.nhaarman.mockitokotlin2.*
-import org.junit.Assert.assertNull
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.mockito.InjectMocks
-import org.mockito.MockitoAnnotations
-import org.mockito.Spy
-import org.mockito.junit.MockitoJUnitRunner
-import javax.inject.Provider
-
-@RunWith(MockitoJUnitRunner::class)
-class AbbreviationsPageDataSourceFactoryTest {
-
-    @Spy
-    lateinit var providerDataSource: Provider<AbbreviationsPageDataSource>
-
-    @Spy
-    lateinit var sourceLiveData: MutableLiveData<AbbreviationsPageDataSource>
-
-    @InjectMocks
-    lateinit var dataSourceFactory: AbbreviationsPageDataSourceFactory
-
-    @Before
-    fun setUp() {
-        MockitoAnnotations.initMocks(this)
-    }
-
-    @Test
-    fun init_WithoutCreate_ShouldNotHaveDataSource() {
-        verify(dataSourceFactory.sourceLiveData, never())
-        assertNull(dataSourceFactory.sourceLiveData.value)
-    }
-
-    @Test
-    fun init_WithCreate_ShouldHaveDataSource() {
-        doReturn(AbbreviationsPageDataSource(mock(), mock())).whenever(providerDataSource).get()
-        val dataSource = dataSourceFactory.create() as AbbreviationsPageDataSource
-        verify(dataSourceFactory.sourceLiveData).postValue(same(dataSource))
-    }
-
-    @Test
-    fun refreshDataSource_WithValue_ShouldInvalidateData() {
-        val dataSource = mock<AbbreviationsPageDataSource>()
-        doReturn(dataSource).whenever(sourceLiveData).value
-
-        dataSourceFactory.refresh()
-
-        verify(dataSource).invalidate()
-        verify(dataSource, never()).retry()
-    }
-
-    @Test
-    fun refreshDataSource_WithoutValue_ShouldInvalidateData() {
-        val dataSource = mock<AbbreviationsPageDataSource>()
-        doReturn(null).whenever(sourceLiveData).value
-
-        dataSourceFactory.refresh()
-
-        verify(dataSource, never()).invalidate()
-        verify(dataSource, never()).retry()
-    }
-
-    @Test
-    fun retryDataSource_WithValue_ShouldRetryData() {
-        val dataSource = mock<AbbreviationsPageDataSource>()
-        doReturn(dataSource).whenever(sourceLiveData).value
-
-        dataSourceFactory.retry()
-
-        verify(dataSource).retry()
-        verify(dataSource, never()).invalidate()
-    }
-
-    @Test
-    fun retryDataSource_WithoutValue_ShouldRetryData() {
-        val dataSource = mock<AbbreviationsPageDataSource>()
-        doReturn(null).whenever(sourceLiveData).value
-
-        dataSourceFactory.retry()
-
-        verify(dataSource, never()).retry()
-        verify(dataSource, never()).invalidate()
-    }
-}
+// @RunWith(MockitoJUnitRunner::class)
+// class AbbreviationsPageDataSourceFactoryTest {
+//
+//    @Spy
+//    lateinit var providerDataSource: Provider<AbbreviationsPageDataSource>
+//
+//    @Spy
+//    lateinit var sourceLiveData: MutableLiveData<AbbreviationsPageDataSource>
+//
+//    @InjectMocks
+//    lateinit var dataSourceFactory: AbbreviationsPageDataSourceFactory
+//
+//    @Before
+//    fun setUp() {
+//        MockitoAnnotations.initMocks(this)
+//    }
+//
+//    @Test
+//    fun init_WithoutCreate_ShouldNotHaveDataSource() {
+//        verify(dataSourceFactory.sourceLiveData, never())
+//        assertNull(dataSourceFactory.sourceLiveData.value)
+//    }
+//
+//    @Test
+//    fun init_WithCreate_ShouldHaveDataSource() {
+//        doReturn(AbbreviationsPageDataSource(mock(), mock())).whenever(providerDataSource).get()
+//        val dataSource = dataSourceFactory.create() as AbbreviationsPageDataSource
+//        verify(dataSourceFactory.sourceLiveData).postValue(same(dataSource))
+//    }
+//
+//    @Test
+//    fun refreshDataSource_WithValue_ShouldInvalidateData() {
+//        val dataSource = mock<AbbreviationsPageDataSource>()
+//        doReturn(dataSource).whenever(sourceLiveData).value
+//
+//        dataSourceFactory.refresh()
+//
+//        verify(dataSource).invalidate()
+//        verify(dataSource, never()).retry()
+//    }
+//
+//    @Test
+//    fun refreshDataSource_WithoutValue_ShouldInvalidateData() {
+//        val dataSource = mock<AbbreviationsPageDataSource>()
+//        doReturn(null).whenever(sourceLiveData).value
+//
+//        dataSourceFactory.refresh()
+//
+//        verify(dataSource, never()).invalidate()
+//        verify(dataSource, never()).retry()
+//    }
+//
+//    @Test
+//    fun retryDataSource_WithValue_ShouldRetryData() {
+//        val dataSource = mock<AbbreviationsPageDataSource>()
+//        doReturn(dataSource).whenever(sourceLiveData).value
+//
+//        dataSourceFactory.retry()
+//
+//        verify(dataSource).retry()
+//        verify(dataSource, never()).invalidate()
+//    }
+//
+//    @Test
+//    fun retryDataSource_WithoutValue_ShouldRetryData() {
+//        val dataSource = mock<AbbreviationsPageDataSource>()
+//        doReturn(null).whenever(sourceLiveData).value
+//
+//        dataSourceFactory.retry()
+//
+//        verify(dataSource, never()).retry()
+//        verify(dataSource, never()).invalidate()
+//    }
+// }
